refactor: missed a test
