+++ tests/test_regression.py
@@ -6,8 +6,6 @@ import pytest
 from . import utils
 from .conftest import FuzzyCallChecker
 from .deployment_utils import UniversumRunner
-from .perforce_utils import P4TestEnvironment
-from .test_p4_regression import perforce_environment
 from .utils import LocalTestEnvironment
 
 
@@ -82,23 +80,3 @@ from universum.configuration_support import Configuration
 configs = Configuration([dict(name="Test configuration", command=["ls", "-la"])])
 """, vcs_type="none", environment=['LANG=en_US', 'LC_ALL=en_US'])
     assert "\u2514" not in output
-
-
-def test_exit_code_failed_report(perforce_environment: P4TestEnvironment):
-    """
-    This test checks for previous bug where exceptions during result reporting led to exit code 0
-    even when '--fail-unsuccessful' option was enabled (and some steps failed)
-    """
-    config = """
-from universum.configuration_support import Configuration
-
-configs = Configuration([dict(name="Unsuccessful step", command=["exit", "1"])])
-"""
-    perforce_environment.shelve_config(config)
-    perforce_environment.settings.MainVcs.report_to_review = True
-    perforce_environment.settings.Swarm.server_url = "some_server"
-    perforce_environment.settings.Swarm.review_id = "some_id"
-    perforce_environment.settings.Swarm.change = perforce_environment.settings.PerforceMainVcs.shelve_cls[0]
-    perforce_environment.settings.Main.fail_unsuccessful = True
-
-    perforce_environment.run(expect_failure=True)
